refactors of directoris
