+++ lib/features/overview/views/OrdersListScreen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:pos_dashboard_v1/features/dashboard/views/desktop_layout/models/order_model.dart';
+import 'package:pos_dashboard_v1/features/overview/models/order_model.dart';
 import 'package:pdf/widgets.dart' as pw;
 import 'package:printing/printing.dart';
-import '../../../../../l10n/app_localizations.dart';
-import '../views/order_details_screen.dart';
+import '../../../l10n/app_localizations.dart';
+import 'order_details_screen.dart';
 
 class OrderListScreen extends StatelessWidget {
   final List<Order> orders;
