+++ lib/features/overview/views/order_details_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:pdf/widgets.dart' as pw;
-import 'package:pos_dashboard_v1/features/dashboard/views/desktop_layout/models/order_model.dart';
+import 'package:pos_dashboard_v1/features/overview/models/order_model.dart';
 import 'package:printing/printing.dart';
 
 class OrderDetailsScreen extends StatelessWidget {
