+++ lib/features/overview/widgets/user_info_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:pos_dashboard_v1/core/utils/manager.dart';
-import 'package:pos_dashboard_v1/features/dashboard/views/desktop_layout/widgets/label_info_card.dart';
+import 'package:pos_dashboard_v1/features/overview/widgets/label_info_card.dart';
 
 // ignore: camel_case_types
 class user_info_card extends StatelessWidget {
