+++ lib/features/overview/widgets/user_info_section.dart
@@ -67,10 +67,10 @@
 // }
 
 import 'package:flutter/material.dart';
-import 'package:pos_dashboard_v1/features/dashboard/views/desktop_layout/widgets/custom_label.dart';
+import 'package:pos_dashboard_v1/features/overview/widgets/custom_label.dart';
 import 'package:pos_dashboard_v1/l10n/app_localizations.dart';
 
-import '../../../../../core/utils/manager.dart';
+import '../../../core/utils/manager.dart';
 
 class UserInfoSection extends StatelessWidget {
   const UserInfoSection({super.key});
